+++ scripts/DiamondSizes.js
@@ -1,4 +1,6 @@
-import { getSizes, setSize } from "./database.js"
+import { getSizes, setSize } from "./dataAccess.js"
+import { dispatchOrderBtnEvent  } from "./orderBtnEvent.js"
+import { checkOrderState } from "./dataAccess.js"
 
 const sizes = getSizes()
 
@@ -7,10 +9,12 @@ document.addEventListener(
     (event) => {
         if (event.target.name === "size") {
           setSize(parseInt(event.target.value))
+          if ( checkOrderState() ) {
+            dispatchOrderBtnEvent()
         }
     }
+}
 )
-
 export const DiamondSizes = () => {
     let html = "<ul>"
