+++ scripts/Metals.js
@@ -1,4 +1,6 @@
-import { getMetals, setMetal } from "./database.js"
+import { getMetals, setMetal } from "./dataAccess.js"
+import { dispatchOrderBtnEvent  } from "./orderBtnEvent.js"
+import { checkOrderState } from "./dataAccess.js"
 
 const metals = getMetals()
 
@@ -7,7 +9,10 @@ document.addEventListener(
     (event) => { 
         if (event.target.name === "metal") {
             setMetal(parseInt(event.target.value))
+            if ( checkOrderState() ) {
+                dispatchOrderBtnEvent()
         }
+    }
     }
 )
 
@@ -15,12 +20,12 @@ export const Metals = () => {
     let html = "<ul>"
 
     // This is how you have been converting objects to <li> elements
-    for (const metal of metals) {
-        html += `<li>
-            <input type="radio" name="metal" value="${metal.id}" /> ${metal.metal}
+    const listItems = metals.map(metal => {
+        return `<li>
+            <input type="radio" name="metal" value="${metal.id}"/> ${metal.metal}
         </li>`
-    }
-
+    })
+    html += listItems.join("")
     html += "</ul>"
     return html
 }
