+++ scripts/database.js
@@ -5,13 +5,8 @@
     modules to get copies of the state.
 
 */
-const database = {
-    orderBuilder: {
-        metalId: 2,
-        styleId: 3,
-        sizeId: 1,
-        typeId: 1
-    },
+export const database = {
+    orderBuilder: {},
 
     styles: [
         { id: 1, style: "Classic", price: 500 },
@@ -50,56 +45,67 @@ const database = {
     ]
 }
 
-export const getMetals = () => {
-    return database.metals.map(metal => ({...metal}))
-}
+// export const getMetals = () => {
+//     return database.metals.map(metal => ({...metal}))
+// }
 
-export const getSizes = () => {
-    return database.sizes.map(size => ({...size}))
-}
+// export const getSizes = () => {
+//     return database.sizes.map(size => ({...size}))
+// }
 
-export const getStyles = () => {
-    return database.styles.map(style => ({...style}))
-}
+// export const getStyles = () => {
+//     return database.styles.map(style => ({...style}))
+// }
 
-export const getTypes = () => {
-    return database.types.map(type => ({...type}))
-}
+// export const getTypes = () => {
+//     return database.types.map(type => ({...type}))
+// }
 
-export const getOrders = () => {
-    return database.customOrders.map(order => ({...order}))
-}
+// export const getOrders = () => {
+//     return database.customOrders.map(order => ({...order}))
+// }
 
-export const setMetal = (id) => {
-    database.orderBuilder.metalId = id
-}
-export const setSize = (id) => {
-    database.orderBuilder.sizeId = id
-}
-export const setStyle = (id) => {
-    database.orderBuilder.styleId = id
-}
-export const setType = (id) => {
-    database.orderBuilder.typeId = id
-}
+// export const setMetal = (id) => {
+//     database.orderBuilder.metalId = id
+  
+// }
+// export const setSize = (id) => {
+//     database.orderBuilder.sizeId = id
+// } 
+// export const setStyle = (id) => {
+//     database.orderBuilder.styleId = id
+   
+// }
+// export const setType = (id) => {
+//     database.orderBuilder.typeId = id
+// }
 
-export const addCustomOrder = () => {
-    // Copy the current state of user choices
-    const newOrder = {...database.orderBuilder}
+// export const checkOrderState = () => {
+//     return (
+//       "metalId" in database.orderBuilder &&
+//       "sizeId" in database.orderBuilder &&
+//       "styleId" in database.orderBuilder &&
+//       "typeId" in database.orderBuilder
+//       )
+//   }
 
-    // Add a new primary key to the object
-    const lastIndex = database.customOrders.length - 1
-    newOrder.id = database.customOrders[lastIndex].id + 1
+// export const addCustomOrder = () => {
+//     // Copy the current state of user choices
+//     const newOrder = {...database.orderBuilder}
 
-    // Add a timestamp to the order
-    newOrder.timestamp = Date.now()
+//     // Add a new primary key to the object
+//     const lastIndex = database.customOrders.length - 1
+//     newOrder.id = lastIndex >= 0 ? database.customOrders[lastIndex].id + 1 : 1
 
-    // Add the new order object to custom orders state
-    database.customOrders.push(newOrder)
+//     // Add a timestamp to the order
+//     newOrder.timestamp = Date.now()
 
-    // Reset the temporary state for user choices
-    database.orderBuilder = {}
+//     // Add the new order object to custom orders state
+//     database.customOrders.push(newOrder)
 
-    // Broadcast a notification that permanent state has changed
-    document.dispatchEvent(new CustomEvent("stateChanged"))
-}
+//     // Reset the temporary state for user choices
+//     database.orderBuilder = {}
+
+//     // Broadcast a notification that permanent state has changed
+//     document.dispatchEvent(new CustomEvent("stateChanged"))
+// }
