+++ scripts/orderBtnEvent.js
@@ -1,0 +1,1 @@
+export const dispatchOrderBtnEvent = () => document.dispatchEvent(new CustomEvent("showOrderBtn"))
